Add Background to HasStarted = true StartlistEntries
Gray background helps visually distinguish and leads to better UX

Also added logic in CompareTo to update HasStarted in StartlistEntry.
While a bit odd, this allows simply calling Sort on the collection
to makr passed entries as HasStarted = true. When combined with a while
loop that Sorts and re-renders on invertal this adds realtime dynamics
making the component more interactive