Allow Edit to reset NotQualified status
If a participation is deemed NotQualified and is edited in a way that
the condtion for not-qualification is no longer met it should be
considered qualifiable again. I.e. if someone is FTQ based on higher
average speed, but is then edited in such a way that their phase speed
drops bellow the restriciton - they should not be excluded in the
ranking

https://github.com/Not-Endurance/not-timing-system/issues/218