Use ParticipationId instead of full Participation in Ranklist
Using full participation results in a data duplication in persistance.
I.e. the file will contain to identical objects, with different
references. This is a problem if a future update for (Tandem name change
for example) comes through as we will have to update all entities where
it's referenced.

When using Document storage (such as our current json files) we should
use Ids instead of multiple references to copies of the same objects

Also this commit adds private constructors with Id parameter in order to
prevent DomainEntity to generate new id on every deserialization. I.e.
to actually persist the ids