Create PrintableComponent as base class instead of BlazerPrinter
This approach is simpler and results in a better UX, because the user
does not have to navigate to a dedicated print page and press another
print button. The tradeoff is that Print has to coordinate with other
Layout components (NotPage as of time of commit) in order to hide
unncessary elements from view. It isn't a big tradeoff however, because
the BlazorRenderer approach also required coordination with MainLayout
anyway

https://github.com/Not-Endurance/not-timing-system/issues/288