Change RanklistPage to printable component
Also fix OnParameterSet in RanklistRow. Async lifecycle hooks aren't
guaranteed to be executed before the intial render is complete. This
caused a problem when printing the Ranklist because at the time the
print engine gets a hold of the HTML content not all entires are yet
rendered and thus are ommited from the print page. Changing
to OnParametersSet fixes this because the method invocation is
synchronous and always completes before the next operation (in this case
window.print)

https://github.com/Not-Endurance/not-timing-system/issues/288