Fix SetupState properties odrder and add explanation
The order here is very important due to how EntityReferenceEqualityGuardConverteer works. Root level entities (Loop, Horse etc) MUST precede their parent entities (Combination, Phase) in order to be the first to be serialized. Otherwise updates on those entities will be ignored because the obsoleted entities will be serialized first and the updates will be lost to a $domainRef

https://github.com/Not-Endurance/not-timing-system/issues/301A